requirements.txt
# Core App
streamlit==1.49.1
openai>=1.0.0
pinecone>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# NLP + Embeddings
sentence-transformers==2.7.0
tiktoken>=0.5.0
torch>=2.1.0
textblob==0.19.0

# ML + Vector ops
numpy>=1.24
scikit-learn>=1.4.0

# Document processing
pypdf>=4.0.0
pdfminer.six>=20221105
python-docx>=1.1.0
pandas>=2.0.0

# Optional Windows extras
pywin32>=306 ; sys_platform == 'win32'
send2trash>=1.8.0 ; sys_platform == 'win32'

#progress bar
tqdm==4.67.1

#Auth
msal==1.34.0  
